WestUtils Standard Update v0.3.9
Description:
Some serious changes to the way the system works. Removed
FormattedStrings and combined their functionality into the
ColoredStrings. Added image capabilities to ColoredStrings. Added a new
formatting package and basic classes. These are just general use
utilities designed for WestGraphics to be more efficiently used.
Essentially, with these formatting classes complex code can be reduced
to just a few lines.

Changes:
 - Fixed #33
 - Fixed #34
 - Fixed #35
 - Modified FXFontMetrics to find the width and height of a
ColoredString as a static function instead of a class function. This
just made more overall sense

Known Issues:
 - Stacked functionality is not supported at this time. This means
stacked drag and drops and stacked scrollable panes will not work the
way they are supposed to

+++ westUtilities/src/graphicsUtilities/WGTheme.java
@@ -63,11 +63,11 @@ public class WGTheme implements TextStyles
     private Paint barColor;
     private Font textFont;
     private HashTable gradientOrientationPreferences;
-    public WGTheme(float borderSize, int textStyle, int textPosition, double textXSizePercent, double textYSizePercent, Paint backgroundColor, Paint borderColor, Paint textColor, Paint scrollBarColor, Paint cursorColor, Paint highlightColor, Paint barColor, Paint checkColor, Font textFont)
+    public WGTheme(double borderSize, int textStyle, int textPosition, double textXSizePercent, double textYSizePercent, Paint backgroundColor, Paint borderColor, Paint textColor, Paint scrollBarColor, Paint cursorColor, Paint highlightColor, Paint barColor, Paint checkColor, Font textFont)
     {
         this(borderSize, textStyle, textPosition, textXSizePercent, textYSizePercent, backgroundColor, borderColor, textColor, scrollBarColor, cursorColor, highlightColor, barColor, checkColor, textFont, DEFAULT_PREFERRENCES);
     }
-    public WGTheme(float borderSize, int textStyle, int textPosition, double textXSizePercent, double textYSizePercent, Paint backgroundColor, Paint borderColor, Paint textColor, Paint scrollBarColor, Paint cursorColor, Paint highlightColor, Paint barColor, Paint checkColor, Font textFont, HashTable gradientOrientationPreferences)
+    public WGTheme(double borderSize, int textStyle, int textPosition, double textXSizePercent, double textYSizePercent, Paint backgroundColor, Paint borderColor, Paint textColor, Paint scrollBarColor, Paint cursorColor, Paint highlightColor, Paint barColor, Paint checkColor, Font textFont, HashTable gradientOrientationPreferences)
     {
         this.textFont = textFont;
         this.textPosition = textPosition;
@@ -202,7 +202,7 @@ public class WGTheme implements TextStyles
     
     
     //Setters:
-    public void setBorderSize(float borderSize) {
+    public void setBorderSize(double borderSize) {
         this.borderSize = borderSize;
     }
